+++ Data structure impl/queue/Dequeue.cs
